Moved extractElFork inside of the ElForkAwareBlockImporter
